+++ src/api/interfaces/pallet_balances.rs
@@ -11,29 +11,15 @@
    limitations under the License.
 */
 use crate::{
-	api::{interfaces::generic_storage::GetGenericStorage, Api, ApiClientError, ApiResult},
+	api::{interfaces::storage::GetStorage, Api, ApiResult},
 	rpc::Request,
 	ExtrinsicParams,
 };
-use ac_compose_macros::rpc_params;
-use ac_primitives::{BalancesConfig, FeeDetails, InclusionFee, RuntimeDispatchInfo};
-use sp_rpc::number::NumberOrHex;
+use ac_primitives::BalancesConfig;
 
 /// Interface to common calls of the substrate balances pallet.
 pub trait GetBalance<Hash> {
 	type Balance;
-
-	fn get_fee_details(
-		&self,
-		xthex_prefixed: &str,
-		at_block: Option<Hash>,
-	) -> ApiResult<Option<FeeDetails<Self::Balance>>>;
-
-	fn get_payment_info(
-		&self,
-		xthex_prefixed: &str,
-		at_block: Option<Hash>,
-	) -> ApiResult<Option<RuntimeDispatchInfo<Self::Balance>>>;
 
 	fn get_existential_deposit(&self) -> ApiResult<Self::Balance>;
 }
@@ -44,63 +30,10 @@ where
 	Client: Request,
 	Runtime: BalancesConfig,
 	Params: ExtrinsicParams<Runtime::Index, Runtime::Hash>,
-	Runtime::Balance: TryFrom<NumberOrHex>,
 {
 	type Balance = Runtime::Balance;
-
-	fn get_fee_details(
-		&self,
-		xthex_prefixed: &str,
-		at_block: Option<Runtime::Hash>,
-	) -> ApiResult<Option<FeeDetails<Self::Balance>>> {
-		let details: Option<FeeDetails<NumberOrHex>> = self
-			.client()
-			.request("payment_queryFeeDetails", rpc_params![xthex_prefixed, at_block])?;
-
-		let details = match details {
-			Some(details) => Some(convert_fee_details(details)?),
-			None => None,
-		};
-		Ok(details)
-	}
-
-	fn get_payment_info(
-		&self,
-		xthex_prefixed: &str,
-		at_block: Option<Runtime::Hash>,
-	) -> ApiResult<Option<RuntimeDispatchInfo<Self::Balance>>> {
-		let res = self
-			.client()
-			.request("payment_queryInfo", rpc_params![xthex_prefixed, at_block])?;
-		Ok(res)
-	}
 
 	fn get_existential_deposit(&self) -> ApiResult<Self::Balance> {
 		self.get_constant("Balances", "ExistentialDeposit")
 	}
-}
-
-fn convert_fee_details<Balance: TryFrom<NumberOrHex>>(
-	details: FeeDetails<NumberOrHex>,
-) -> ApiResult<FeeDetails<Balance>> {
-	let inclusion_fee = if let Some(inclusion_fee) = details.inclusion_fee {
-		Some(inclusion_fee_with_balance(inclusion_fee)?)
-	} else {
-		None
-	};
-	let tip = details.tip.try_into().map_err(|_| ApiClientError::TryFromIntError)?;
-	Ok(FeeDetails { inclusion_fee, tip })
-}
-
-fn inclusion_fee_with_balance<Balance: TryFrom<NumberOrHex>>(
-	inclusion_fee: InclusionFee<NumberOrHex>,
-) -> ApiResult<InclusionFee<Balance>> {
-	Ok(InclusionFee {
-		base_fee: inclusion_fee.base_fee.try_into().map_err(|_| ApiClientError::TryFromIntError)?,
-		len_fee: inclusion_fee.len_fee.try_into().map_err(|_| ApiClientError::TryFromIntError)?,
-		adjusted_weight_fee: inclusion_fee
-			.adjusted_weight_fee
-			.try_into()
-			.map_err(|_| ApiClientError::TryFromIntError)?,
-	})
 }
